# Conflicts:
#	mobile/app/src/main/res/layout/activity_menu.xml

+++ mobile/app/src/main/java/com/projetosuniso/digdin/requisicoes/conta/ContaAtualiza.java
@@ -16,9 +16,9 @@ import java.net.URL;
 public class ContaAtualiza extends AsyncTask<Void, Void, String> {
 
     private final String conta;
-    private final int id;
+    private final Long id;
 
-    public ContaAtualiza(String conta, int id) {
+    public ContaAtualiza(String conta, Long id) {
         this.conta = conta;
         this.id = id;
     }
